First draft of the proposal
Added a first draft that certainly warrants additional work. Little
research has been performed (on my part) at this point, so it's more of
a "statement of objectives" -- nearly what we discussed in-person.

So, I think the next step is research. Can we clarify the goals we
believe a system should meet? We also need to find "real" references
(not Wikipedia). Finally, we need to clarify what we're going to do, as
I don't think I accurately captured it in what I have written at
present.